Add filtering of delivery types for provider course response

+++ src/SFA.DAS.FindApprenticeshipTraining/InnerApi/Responses/GetDeliveryTypeItem.cs
@@ -3,6 +3,6 @@
     public class GetDeliveryTypeItem
     {
         public string DeliveryModes { get; set; }
-        public double DistanceInMiles { get; set; }
+        public decimal DistanceInMiles { get; set; }
     }
 }
